requirements.txt
h5py==3.9.0
munkres==1.1.4
numpy==1.21.1
pandas==1.4.4
scanpy==1.9.6
scikit_learn==1.1.3
scipy==1.9.3
torch==1.13.1
pyyaml==6.0.2